nexus/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic[email]==2.6.4
python-dotenv==1.0.1
asyncpg==0.29.0
databases==0.9.0
httpx==0.27.0
# AI & Security
google-cloud-aiplatform==1.42.0
openai==2.14.0
cryptography==42.0.5
# Gmail API with OAuth
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0


google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0


google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0


google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

